+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Unreleased
 
 - Add detection for improper keyword use. [#203](https://github.com/BernieWhite/PSRule/issues/203)
+- Automatically load rule modules. [#218](https://github.com/BernieWhite/PSRule/issues/218)
 
 ## v0.8.0-B190716 (pre-release)